backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.3
django-environ==0.13.0
djangorestframework==3.16.1
idna==3.11
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
